feat: SOLID principles - Single Responsibility
